server/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
beanie==1.29.0
captcha==0.7.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
docker==7.1.0
fastapi==0.115.12
h11==0.12.0
httpcore==0.15.0
httptools==0.6.4
httpx==0.23.0
idna==3.10
lazy-model==0.2.0
loguru==0.7.3
lxml==5.4.0
minio==7.2.15
motor==3.7.1
pillow==11.2.1
pip==25.1.1
pycparser==2.22
pycryptodome==3.23.0
pydantic==2.11.5
pydantic-core==2.33.2
pydantic-settings==2.9.1
pyjwt==2.10.1
pymongo==4.13.0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
redis==6.1.0
requests==2.32.3
rfc3986==1.5.0
sniffio==1.3.1
starlette==0.46.2
toml==0.10.2
typing-extensions==4.13.2
typing-inspection==0.4.1
urllib3==2.4.0
uvicorn==0.34.2
watchfiles==1.0.5
websockets==15.0.1
thefuzz==0.22.1
apscheduler==3.10.4
